Added some user auth for making favorites and searching on the API
